Extract queued registration related items from component
